requirements.txt
python-telegram-bot==13.7
python-dotenv==0.19.0
openai==0.27.0